refactor: convert lesson creation from modal to standalone page
